hamr disable CUDA object support
adds a CMake option HAMR_CUDA_OBJECTS which defaults to ON that can be
used to disable CUDA support for memory managment on clasess. When
disabled allocators and copiers abort if used with a class type.  this
is desirable when dealing with classes that do not have CUDA device
constructors and asignment operators (eg: std::string). In that case
it's up to the user to avoid CUDA allocators with classes.