buffer move assignment behavior
this brings the move assignment behavior in line with the documentation.
if the type and allocator are the same the contents are moved, otherwise
the contents are copied.

this behavior puts more control of the location of the data in the
programmers hands. they can now specify the destination location while
benefiting from the low overheads of a move when everything matches.