buffer pointer API
the new API lets one access the std::shared_ptr managing the data
directly. this avoids to costs of the logic that determines when
a temporary copy is needed. it's only safe to use when one knows
where the data resides (much in the same way as the data API)