buffer fix multi-device copy and move construct
the original implementation assumed the new/destination buffer and the
existing/source buffer were on the same device and would have failed if
that was not the case. The original implementation used the allocator,
to allocate and initialize in the same step. the fix allocates and
initialized in two steps. This could be less efficient in some cases.
Alternatively the allocator could be revised to handle data from
multiple devices.