add buffer_util to simplify data access
this utility contains a number of free functions that forward calls to
hamr::buffer<T>::get_X_accessible returning a std::tuple of shared_ptr<const T>
and const T* for each argument. this cuts the number of lines of code needed
for read only access to buffer contents by 1/2.