separate implementation and explicit instantiation
factors declaration and implementation into different files. by default
implementations are included into declaration files but can be excldued
and explicit instantiations compiled into the library. This is needed
with CLang for HIP/CUDA and OpenMP to be used in the same library. A build
variable HAMR_SEPARATE_IMPL controls. It is set when CUDA and OpenMP
are requested in the same build and the compiler is not NVHPC.